+++ src/flow2pu_function.ps1
@@ -185,6 +185,9 @@ function flow2pu {
         [string] $FontName,
 
         [Parameter( Mandatory=$False)]
+        [int] $FontSize,
+
+        [Parameter( Mandatory=$False)]
         [string] $FontNameWindowsDefault = "Meiryo",
 
         [Parameter( Mandatory=$False)]
@@ -452,15 +455,18 @@ function flow2pu {
         if($IsWindows){
             ## case  windows
             if($FontName){
-                $readLineAryHeader += "skinparam DefaultFontName ""$FontName"""
+                $readLineAryHeader += "skinparam defaultFontName ""$FontName"""
             } else {
-                $readLineAryHeader += "skinparam DefaultFontName ""$FontNameWindowsDefault"""
+                $readLineAryHeader += "skinparam defaultFontName ""$FontNameWindowsDefault"""
             }
         } else {
             ## case linux and mac
             if($FontName){
-                $readLineAryHeader += "skinparam DefaultFontName ""$FontName"""
+                $readLineAryHeader += "skinparam defaultFontName ""$FontName"""
             }
+        }
+        if ($FontSize){
+            $readLineAryHeader += "skinparam defaultFontSize $FontSize"
         }
         if ($Monochrome){
             $readLineAryHeader += "skinparam monochrome true"