+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added [ysort], [ycalc], [fval] functions.
 - Added [wrap] function.
 - Added [percentile] supports multiple value fields. Missing/Empty value detection/removal/replacement.
+- Added [flow2pu] [gantt2pu] [logi2pu] [mind2pu] [seq2pu] `-FontSize` option.
 
 
 ## [2.0.0]